Move `Faces` to `shape::faces`
