+++ src/kernel/topology/faces.rs
@@ -16,23 +16,6 @@ use crate::{
 
 use super::edges::Cycle;
 
-/// The faces of a shape
-#[derive(Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
-pub struct Faces(pub Vec<Face>);
-
-impl Faces {
-    pub fn triangles(
-        &self,
-        tolerance: Scalar,
-        out: &mut Vec<Triangle<3>>,
-        debug_info: &mut DebugInfo,
-    ) {
-        for face in &self.0 {
-            face.triangles(tolerance, out, debug_info);
-        }
-    }
-}
-
 /// A face of a shape
 #[derive(Clone, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
 pub enum Face {
